# public/swagger.json
{
    "components": {
        "examples": {},
        "headers": {},
        "parameters": {},
        "requestBodies": {},
        "responses": {},
        "schemas": {
            "IJsonResponseSuccess_any_": {
                "properties": {
                    "status": {
                        "type": "number",
                        "format": "double"
                    },
                    "message": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "data": {}
                },
                "required": [
                    "status"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "IUserParam": {
                "properties": {
                    "id": {
                        "type": "number",
                        "format": "double"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "birthday": {
                        "type": "string"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "IJsonResponseSuccess_IUserParam_": {
                "properties": {
                    "status": {
                        "type": "number",
                        "format": "double"
                    },
                    "message": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/IUserParam"
                    }
                },
                "required": [
                    "status"
                ],
                "type": "object",
                "additionalProperties": false
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "user-management",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Santiago"
        }
    },
    "openapi": "3.0.0",
    "paths": {
        "/users/all": {
            "get": {
                "operationId": "SearchForms",
                "responses": {
                    "200": {
                        "description": "Return array of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IJsonResponseSuccess_any_"
                                }
                            }
                        }
                    }
                },
                "description": "Get all users",
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/users/{id}": {
            "get": {
                "operationId": "GetForm",
                "responses": {
                    "200": {
                        "description": "Return user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IJsonResponseSuccess_IUserParam_"
                                }
                            }
                        }
                    }
                },
                "description": "Get user by id",
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            }
        }
    },
    "servers": [
        {
            "url": "/api"
        }
    ]
}